Created "Cluster" object and added auto duration adjustment
Added logic to cap query duration at cluster age, in case cluster is younger than the requested duration is long. To facilitate this, I change most uses of the datetime object to be "timezone-aware" and created a Cluster object in the uhc module to make information passing easier. Clusters that have their duration modified have an asterisk (*) appended to their names